move monorail and metadata to public
