+++ packages/cli-kit/src/index.ts
@@ -5,5 +5,3 @@ export * as schema from './schema.js'
 export * as store from './store.js'
 export * as ui from './ui.js'
 export * as outputMocker from './testing/output.js'
-export * as metadata from './metadata.js'
-export * as monorail from './monorail.js'
