+++ packages/cli-kit/src/private/node/analytics.ts
@@ -2,7 +2,7 @@ import {hashString} from '../../public/node/crypto.js'
 import {getPackageManager, packageManagerUsedForCreating} from '../../public/node/node-package-manager.js'
 import BaseCommand from '../../public/node/base-command.js'
 import {CommandContent} from '../../public/node/hooks/prerun.js'
-import * as metadata from '../../metadata.js'
+import * as metadata from '../../public/node/metadata.js'
 import {platformAndArch} from '../../public/node/os.js'
 import {Interfaces} from '@oclif/core'
 import {ciPlatform, cloudEnvironment, macAddress} from '@shopify/cli-kit/node/environment/local'
