+++ packages/cli-kit/src/public/node/analytics.ts
@@ -1,8 +1,8 @@
 import {version as rubyVersion} from './ruby.js'
 import {alwaysLogAnalytics, analyticsDisabled, isShopify} from './environment/local.js'
+import * as metadata from './metadata.js'
+import {publishEvent, MONORAIL_COMMAND_TOPIC} from './monorail.js'
 import {content, debug, token} from '../../output.js'
-import * as metadata from '../../metadata.js'
-import {publishEvent, MONORAIL_COMMAND_TOPIC} from '../../monorail.js'
 import {fanoutHooks} from '../../plugins.js'
 import {getEnvironmentData, getSensitiveEnvironmentData} from '../../private/node/analytics.js'
 import {CLI_KIT_VERSION} from '../common/version.js'
