+++ packages/cli-kit/src/public/node/base-command.ts
@@ -1,9 +1,9 @@
 import {errorHandler, registerCleanBugsnagErrorsFromWithinPlugins} from './error-handler.js'
 import {loadPresetsFromDirectory} from './presets.js'
 import {isDevelopment} from './environment/local.js'
+import {addPublic} from './metadata.js'
 import {JsonMap} from '../../private/common/json.js'
 import {Abort} from '../../error.js'
-import {addPublic} from '../../metadata.js'
 import {content, info, token} from '../../output.js'
 import {hashString} from '../../public/node/crypto.js'
 import {isTruthy} from '../../private/node/environment/utilities.js'
