+++ packages/cli-kit/src/public/node/error-handler.ts
@@ -1,5 +1,6 @@
 import {reportAnalyticsEvent} from './analytics.js'
 import * as path from './path.js'
+import * as metadata from './metadata.js'
 import {getEnvironmentData} from '../../private/node/analytics.js'
 import {
   AbortSilent,
@@ -10,7 +11,6 @@ import {
   cleanSingleStackTracePath,
 } from '../../error.js'
 import {debug, info} from '../../output.js'
-import * as metadata from '../../metadata.js'
 import {fanoutHooks} from '../../plugins.js'
 import {bugsnagApiKey} from '../../private/node/constants.js'
 import {CLI_KIT_VERSION} from '../common/version.js'
