+++ packages/cli-kit/src/public/node/metadata.test.ts
@@ -1,10 +1,10 @@
 import {createRuntimeMetadataContainer} from './metadata.js'
-import * as errorHandler from './public/node/error-handler.js'
+import * as errorHandler from './error-handler.js'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
 
 describe('runtime metadata', () => {
   beforeEach(() => {
-    vi.mock('./public/node/error-handler.js')
+    vi.mock('./error-handler.js')
   })
 
   it('can manage data', async () => {
