+++ packages/cli-kit/src/public/node/metadata.ts
@@ -1,8 +1,8 @@
-import {AnyJson} from './private/common/json.js'
 import {MonorailEventPublic} from './monorail.js'
-import {sendErrorToBugsnag} from './public/node/error-handler.js'
-import {PickByPrefix} from './public/common/ts/pick-by-prefix.js'
-import {isUnitTest} from './public/node/environment/local.js'
+import {sendErrorToBugsnag} from './error-handler.js'
+import {isUnitTest} from './environment/local.js'
+import {PickByPrefix} from '../common/ts/pick-by-prefix.js'
+import {AnyJson} from '../../private/common/json.js'
 
 type ProvideMetadata<T> = () => Partial<T> | Promise<Partial<T>>
 
@@ -14,6 +14,11 @@ type MetadataErrorHandling =
   // Errors are not caught and will bubble out as normal
   | 'bubble'
 
+/**
+ * Get the error handling strategy for metadata.
+ *
+ * @returns 'mute-and-report' in production, 'bubble' in tests.
+ */
 function getMetadataErrorHandlingStrategy(): 'mute-and-report' | 'bubble' {
   if (isUnitTest()) {
     return 'bubble'
@@ -22,13 +27,13 @@ function getMetadataErrorHandlingStrategy(): 'mute-and-report' | 'bubble' {
 }
 
 export interface RuntimeMetadataManager<TPublic extends AnyJson, TSensitive extends AnyJson> {
-  /** Add some public metadata -- this should not contain any PII */
+  /** Add some public metadata -- this should not contain any PII. */
   addPublic: (getData: ProvideMetadata<TPublic>, onError?: MetadataErrorHandling) => Promise<void>
-  /** Add some potentially sensitive metadata -- this may include PII, but unnecessary data should never be tracked (this is a good fit for command args for instance) */
+  /** Add some potentially sensitive metadata -- this may include PII, but unnecessary data should never be tracked (this is a good fit for command args for instance). */
   addSensitive: (getData: ProvideMetadata<TSensitive>, onError?: MetadataErrorHandling) => Promise<void>
-  /** Get a snapshot of the tracked public data */
+  /** Get a snapshot of the tracked public data. */
   getAllPublic: () => Partial<TPublic>
-  /** Get a snapshot of the tracked sensitive data */
+  /** Get a snapshot of the tracked sensitive data. */
   getAllSensitive: () => Partial<TSensitive>
 }
 
@@ -37,9 +42,9 @@ export type SensitiveSchema<T> = T extends RuntimeMetadataManager<infer _TPublic
 
 /**
  * Creates a container for metadata collected at runtime.
- *
  * The container provides async-safe functions for extracting the gathered metadata, and for setting it.
  *
+ * @returns A container for the metadata.
  */
 export function createRuntimeMetadataContainer<
   TPublic extends AnyJson,
